mcp-google-calendar/pyproject.toml
[project]
name = "mcp-google-calendar"
version = "0.1.0"
description = "MCP сервер для работы с Google Calendar через Service Account"
authors = []
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastmcp>=2.10.0,<3.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.0",
    "google-auth>=2.0.0",
    "google-api-python-client>=2.0.0",
    "pytz>=2024.1",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]






